+++ example/lib/pages/credit_card_format_page.dart
@@ -22,11 +22,6 @@ class _CreditCardFormatPageState extends State<CreditCardFormatPage> {
 
   @override
   Widget build(BuildContext context) {
-    final negativeString = toCurrencyString(
-      '-888',
-      mantissaLength: 2,
-    );
-    print(negativeString);
     return Unfocuser(
       child: Scaffold(
         appBar: AppBar(