+++ lib/formatters/money_input_formatter.dart
@@ -26,7 +26,6 @@ THE SOFTWARE.
 
 import 'dart:math';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 import 'formatter_utils.dart';
@@ -45,6 +44,7 @@ class MoneySymbols {
   static const String BITCOIN_SIGN = 'Ƀ';
   static const String SWISS_FRANK_SIGN = '₣';
   static const String RUBLE_SIGN = '₽';
+  static const String WON_SIGN = '₩';
 }
 
 @Deprecated(
@@ -72,7 +72,7 @@ class MoneyInputFormatter extends TextInputFormatter {
   /// added in front of the resulting string. E.g. $ or €
   /// some of the signs are available via constants like [MoneySymbols.EURO_SIGN]
   /// but you can basically add any string instead of it. The main rule is that the string
-  /// must not contain digits, preiods, commas and dashes
+  /// must not contain digits, periods, commas and dashes
   /// [trailingSymbol] is the same as leading but this symbol will be added at the
   /// end of your resulting string like 1,250€ instead of €1,250
   /// [useSymbolPadding] adds a space between the number and trailing / leading symbols